requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0
tflite-runtime==2.14.0
